Issue #96 - hook up geoservice client in gedbrowser
All of the geo information is now in GeoJSON forms. The internal geometry
description is now a Point (location) and 2 Features (bounds and viewport).

The next step will be to make it be completely a Feature Collection. That
will require turning the Point into a Feature with the Point and the
location type (as a property).