Issue #96 - Backed off a little on the GeoJSON
All of the response data is now in the GeoJSON. However, I can't
deserialize the Google data types if they are in a Feature.properties.
So, I have that data redundantly in the response. For current semantics
it's in the wrapper.

Next steps is to make componentize the client. Use the application just
to test the component. Eventually use that component in GedBrowser.