Resolves #181 -	selenium test of development
Parameterized the host a port of the selenium tests so that
they can readily be run against different instances depending
on branch. That also allows running in Eclipse against the
application also running in Eclipse.

I created the new job to run Selenium tests against the
development branch. That requires the paramterization of host
and port.

I added the port definition to the Selenium job on the master
branch. That will have no impact until such time as these
source changes are merged to master.

I updated the readme to point to the new Jenkins job for
running Selenium against the development branch.