Fixes #161 - Deprecate 2 argument insert in Root
The title aspect of change unifies the interface so that Root is
the same as all GedObjects as far as insert is concerned.

Together with this I eliminated the constructors in the data
model that only take a parent. Most of these were being used either:

* in places where a string was being added after construction
* or in being passed a null

Switched the null versions to using a default constructor. Switched
the instance where a string was being set immediately after to use
the version with a parent and a string.