Fixes #146 - remove instanceof
This first round removes the getDate, getBirthDate, etc. methods.
They are replaced with using a visitor that is good for getting
dates out of the datamodel.

This will be the first of several commits that build up this this
pull request until we have eliminated all use of instanceof in
loops through the datamodel.