Fixes #385 - source in header isn't a link
The source item in a GEDCOM header isn't a link. It is a block
describing the source software of that particular GEDCOM file.
As such it needs to be treated as an attribute of the header
and not as a link or as a source object.

Change the object factory used when parsing GEDCOM files to
reflect that changed thinking.