Changed from primeng dialogs to material dialogs
This changes the management of the selection model in the list
pickers. This will have to be cleaned up, probably toward always
using an emitter approach.

There are still remaining problems with dialog geometry management.
In particular, the lists get wide and the header doesn't match the
width of the dialog (the width of the scroll bar seems to be an
issue).

The population and responsiveness of the material selection list is
pretty slow. We'll have to see what to do about that. I don't think
it's a server response time issue, I think it's in the Javascript.

The remaining primeng pieces are the main list pages (persons, notes,
sources...) and the reordering support.